errors

+++ controllers/replyController.js
@@ -3,20 +3,30 @@ const Comment = require("../models/Comment");
 const User = require("../models/User");
 
 const createReply = async (req, res) => {
-  const { userId, text } = req.body;
+  const {  text } = req.body;
   const {commentId} = req.params;
+  const userId= req.user.user_id;
 
   try {
     const user = await User.findById(userId);
     if (!user) {
       return res.status(400).json({ error: "User not found" });
     }
-
+    const username = user.username;
     const comment = await Comment.findById(commentId);
     if (!comment) {
       return res.status(400).json({ error: "Comment not found" });
     }
-    const reply = new Reply({ userId, text });
+    const today = new Date();
+    today.setHours(0, 0, 0, 0);
+    const replyCount = await Reply.countDocuments({
+      userId: userId,
+      createdAt: { $gte: today },
+    });
+    if (replyCount >= 3) {
+      return res.status(400).json({ error: "You have reached the daily reply limit" });
+    }
+    const reply = new Reply({ userId, username, text });
     await reply.save();
 
     comment.replies.push(reply);
@@ -37,13 +47,8 @@ const getReplies = async (req, res) => {
     if (!comment) {
       return res.status(400).json({ error: "Comment not found" });
     }
-    const replies = [];
-    for (const replyId of comment.replies) {
-      const reply = await Reply.findById(replyId);
-      if (reply) {
-        replies.push(reply);
-      }
-    }
+    const replies = await Reply.find({ _id: { $in: comment.replies } }).sort({ editedAt: -1 });
+    
     res.json(replies);
   } catch (error) {
     console.error(error);
@@ -54,13 +59,16 @@ const getReplies = async (req, res) => {
 const updateReply = async (req, res) => {
   const { replyId } = req.params;
   const { text } = req.body;
+  const userId= req.user.user_id;
 
   try {
     const reply = await Reply.findById(replyId);
     if (!reply) {
       return res.status(404).json({ error: "Reply not found" });
     }
-
+    if(userId!==reply.userId.toString()){
+      return res.status(401).json({error:"Cannot edit reply"});
+    }
     reply.text = text;
     reply.editedAt = Date.now();
     await reply.save();
@@ -74,13 +82,16 @@ const updateReply = async (req, res) => {
 
 const deleteReply = async (req, res) => {
   const { replyId } = req.params;
+  const userId= req.user.user_id;
 
   try {
     const reply = await Reply.findById(replyId);
     if (!reply) {
       return res.status(404).json({ error: "Reply not found" });
     }
-
+    if(userId!==reply.userId.toString()){
+      return res.status(401).json({error:"Cannot delete reply"});
+    }
     await reply.deleteOne();
 
     res.status(200).json({ message: "Reply deleted successfully" });
